neural_processing/requirements_nn.txt
insightface==0.7.3
loguru==0.7.0
onnxruntime==1.15.1
opennsfw2==0.10.2
Pillow==9.5.0
psutil==5.9.0
pydantic==1.10.9
python-dotenv==1.0.0
redis==4.6.0
tensorflow==2.12.0
torch==2.0.1
tqdm==4.65.0